+++ exp/services/recoverysigner/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/google/tink/go/hybrid"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/stellar/go/exp/services/recoverysigner/cmd"
@@ -21,7 +22,10 @@ func main() {
 
 	rootCmd.AddCommand((&cmd.ServeCommand{Logger: logger}).Command())
 	rootCmd.AddCommand((&cmd.DBCommand{Logger: logger}).Command())
-	rootCmd.AddCommand((&cmd.KeysetCommand{Logger: logger}).Command())
+
+	// Key template used for the generation of new keys and keysets.
+	keyTemplate := hybrid.ECIESHKDFAES128GCMKeyTemplate()
+	rootCmd.AddCommand((&cmd.KeysetCommand{Logger: logger, KeyTemplate: keyTemplate}).Command())
 
 	err := rootCmd.Execute()
 	if err != nil {
